refactor: split account view model into accountNotificationsIndexViewModel
